--- a/src/Interpreters/RowRefs.cpp
+++ b/src/Interpreters/RowRefs.cpp
@@ -74,9 +74,8 @@ class SortedLookupVector : public SortedLookupVectorBase
 
 
 public:
-    using Keys = std::vector<TKey>;
-    using Entries = PaddedPODArray<Entry>;
-    using RowRefs = PaddedPODArray<RowRef>;
+    using Entries = PODArrayWithStackMemory<Entry, sizeof(Entry)>;
+    using RowRefs = PODArrayWithStackMemory<RowRef, sizeof(RowRef)>;
 
     static constexpr bool is_descending = (inequality == ASOFJoinInequality::Greater || inequality == ASOFJoinInequality::GreaterOrEquals);
     static constexpr bool is_strict = (inequality == ASOFJoinInequality::Less) || (inequality == ASOFJoinInequality::Greater);