--- a/tests/requirements.txt
+++ b/tests/requirements.txt
@@ -1,2 +1,2 @@
-Jinja2==3.1.3
+Jinja2==3.1.4
 termcolor==2.0.1