@yokofly 2024-05-09
as the code changes, in my local testing, if the input data volume is huge, the performance improvement is small.
```
SELECT
  COUNT(*) AS count
FROM
  (
    SELECT
      number % 1000 AS visitor_id, number AS id
    FROM
      system.numbers
    LIMIT 100000000
  ) AS sessions
ASOF LEFT JOIN (
    SELECT
      number % 1000000 AS visitor_id, number AS starting_session_id
    FROM
      system.numbers
    LIMIT 100000000
  ) AS visitors ON (visitors.visitor_id = sessions.visitor_id) AND (visitors.starting_session_id <= sessions.id)

Query id: 054a7265-de5c-4cd5-a888-5ca7cf355d14

┌─────count─┐
│ 100000000 │
└───────────┘

1 row in set. Elapsed: 9.436 sec. Processed 200.02 million rows, 1.60 GB (21.20 million rows/s., 169.57 MB/s.)

c1 :) 
```