@yokofly 2024-05-10
the related code change below:
https://github.com/timeplus-io/proton/blob/8423708d92a01d2682ef5bc03ec28b360a852808/src/Interpreters/HashJoin.cpp#L514-L524
https://github.com/timeplus-io/proton/blob/8423708d92a01d2682ef5bc03ec28b360a852808/src/Interpreters/RowRefs.cpp#L83-L93
https://github.com/timeplus-io/proton/blob/8423708d92a01d2682ef5bc03ec28b360a852808/src/Common/Allocator.h#L301-L330