@yokofly 2024-05-10
in this special case (demo SQL), in the  `AllocatorWithStackMemory::alloc(size)`  `size = 16 initial_bytes = 16`,  and each time the insert entries size is 1, ensure the effective memory usage in stack pure array.

for other case, this alloc will fallback to the formal Base::alloc which will introduce extra overhead.